pyproject.toml
[tool.poetry]
name = "product-testing"
version = "0.1.0"
description = ""
authors = ["dlee <dlee9997@gmail.com>"]
readme = "README.md"
packages = [{include = "product_testing"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
facebook-business = "^18.0.4"
facebookads = "^2.11.4"
cloudflare = "^2.12.4"
llama-cpp-python = "^0.2.18"
python-dotenv = "^1.0.0"
openai = "^1.3.3"
langchain = "0.0.330"
rich = "^13.7.0"
shopifyapi = "^12.3.0"
dropbox = "^11.36.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
